Search functionality to filter asteroids by name.
