+++ src/constants.js
@@ -3,11 +3,10 @@ export const START_DATE_ID = "start-date";
 export const END_DATE_ID = "end-date";
 export const CARD_CONTAINER_ID = "card-container";
 export const HEADER_ID = "header";
-export const FOOTER_ID = "footer";
 
 export const CARDS_PER_PAGE = 6;
+export const TIMEOUT = 5000;
 export const SELECTOR_ID = "selector";
-export const DATE_CONTAINER_ID = "date-container";
 export const BTN_CONTAINER_ID = "btn-container";
 export const IMG_CONTAINER = "img-container";
 export const INTRO_CONTAINER = "intro-container";
@@ -17,10 +16,8 @@ export const SCROLL_TH = 60;
 export const IMG_DAY_ID = "img-day";
 
 //sections
-
 export const INTRO_SECTION_ID = "intro-section";
 export const NEO_SECTION_ID = "neo-section";
-export const VIS_SECTION_ID = "vis-section";
 export const IMG_SECTION_ID = "img-section";
 export const TO_TOP_BTN_ID = "top";
 export const SEARCH_INPUT_ID = "search-input";
