--- src/vis.js
@@ -1,2 +1,0 @@
-
-
